cl: Introduce deps
Useful for dependencies
