+++ src/v/cluster_link/deps.h
@@ -1,0 +1,63 @@
+/*
+ * Copyright 2025 Redpanda Data, Inc.
+ *
+ * Use of this software is governed by the Business Source License
+ * included in the file licenses/BSL.md
+ *
+ * As of the Change Date specified in that file, in accordance with
+ * the Business Source License, use of this software will be governed
+ * by the Apache License, Version 2.0
+ */
+
+#pragma once
+
+#include "cluster_link/fwd.h"
+#include "cluster_link/model/types.h"
+#include "kafka/data/rpc/deps.h"
+#include "model/fundamental.h"
+
+namespace cluster_link {
+
+/**
+ * @brief Abstract class that provides accessors to cluster link table
+ *
+ */
+class link_registry {
+public:
+    link_registry() = default;
+    link_registry(const link_registry&) = delete;
+    link_registry(link_registry&&) = delete;
+    link_registry& operator=(const link_registry&) = delete;
+    link_registry& operator=(link_registry&&) = delete;
+    virtual ~link_registry() = default;
+
+    virtual std::optional<std::reference_wrapper<const model::metadata>>
+      find_link_by_id(model::id_t) const = 0;
+
+    virtual std::optional<std::reference_wrapper<const model::metadata>>
+    find_link_by_name(const model::name_t&) const = 0;
+
+    virtual chunked_vector<model::id_t> get_all_link_ids() const = 0;
+};
+
+/**
+ * @brief Factory abstract class to create new links
+ *
+ */
+class link_factory {
+public:
+    link_factory() = default;
+    link_factory(const link_factory&) = delete;
+    link_factory(link_factory&&) = delete;
+    link_factory& operator=(const link_factory&) = delete;
+    link_factory& operator=(link_factory&&) = delete;
+    virtual ~link_factory() = default;
+
+    virtual std::unique_ptr<link> create_link(
+      ::model::node_id self,
+      model::metadata config,
+      kafka::data::rpc::partition_leader_cache* partition_leader_cache,
+      kafka::data::rpc::partition_manager* partition_manager)
+      = 0;
+};
+} // namespace cluster_link
