+++ src/v/cluster_link/manager.h
@@ -12,6 +12,7 @@
 #pragma once
 
 #include "absl/container/flat_hash_map.h"
+#include "cluster_link/deps.h"
 #include "cluster_link/link.h"
 #include "cluster_link/model/types.h"
 #include "cluster_link/task.h"
@@ -22,49 +23,6 @@
 #include "utils/mutex.h"
 
 namespace cluster_link {
-
-/**
- * @brief Abstract class that provides accessors to cluster link table
- *
- */
-class link_registry {
-public:
-    link_registry() = default;
-    link_registry(const link_registry&) = delete;
-    link_registry(link_registry&&) = delete;
-    link_registry& operator=(const link_registry&) = delete;
-    link_registry& operator=(link_registry&&) = delete;
-    virtual ~link_registry() = default;
-
-    virtual std::optional<std::reference_wrapper<const model::metadata>>
-      find_link_by_id(model::id_t) const = 0;
-
-    virtual std::optional<std::reference_wrapper<const model::metadata>>
-    find_link_by_name(const model::name_t&) const = 0;
-
-    virtual chunked_vector<model::id_t> get_all_link_ids() const = 0;
-};
-
-/**
- * @brief Factory abstract class to create new links
- *
- */
-class link_factory {
-public:
-    link_factory() = default;
-    link_factory(const link_factory&) = delete;
-    link_factory(link_factory&&) = delete;
-    link_factory& operator=(const link_factory&) = delete;
-    link_factory& operator=(link_factory&&) = delete;
-    virtual ~link_factory() = default;
-
-    virtual std::unique_ptr<link> create_link(
-      ::model::node_id self,
-      model::metadata config,
-      kafka::data::rpc::partition_leader_cache* partition_leader_cache,
-      kafka::data::rpc::partition_manager* partition_manager)
-      = 0;
-};
 
 /**
  * @brief Class used to manage cluster links
